requirements.txt
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.2
decorator==5.1.1
emoji==0.5.4
future==0.18.3
pycparser==2.21
python-telegram-bot==12.7
tornado==6.3.2
requests==2.31.0
python-dotenv==1.0.0